split out adjuster logic
