+++ src/pvnet_app/consts.py
@@ -4,3 +4,5 @@ nwp_ukv_path = "nwp_ukv.zarr"
 nwp_ecmwf_path = "nwp_ecmwf.zarr"
 nwp_cloudcasting_path = "nwp_cloudcasting.zarr"
 generation_path = "generation.zarr"
+
+forecast_version = "2.8.0"
