+++ src/pvnet_app/save.py
@@ -2,7 +2,6 @@
 
 import asyncio
 import logging
-from datetime import datetime
 
 import fsspec
 import numpy as np
@@ -10,16 +9,14 @@ import pandas as pd
 import xarray as xr
 from betterproto.lib.google.protobuf import Struct, Value
 from ocf import dp
+from pvnet_app.adjuster import calculate_adjusted_forecast
+from pvnet_app.utils import convert_to_utc_datetime
+from pvnet_app.consts import forecast_version
 
 logger = logging.getLogger(__name__)
 
 
 DATAPLATFORM_MAX_VALUE = 1.09
-
-
-def convert_to_utc_datetime(ts: pd.Timestamp) -> datetime:
-    """Converts internal naive UTC timestamp to aware UTC datetime."""
-    return ts.tz_localize("UTC").to_pydatetime()
 
 
 async def fetch_locations(
@@ -68,7 +65,7 @@ async def fetch_locations(
 def extract_location_capacities_mwp(
     locations: dict[int, dp.ListLocationsResponseLocationSummary],
 ) -> dict[int, float]:
-    """Extract capacities from location summaries."""
+    """Extract capacities in MW from location summaries."""
     return {gsp_id: loc.effective_capacity_watts / 1e6 for gsp_id, loc in locations.items()}
 
 
@@ -78,10 +75,7 @@ async def fetch_or_create_forecaster(
 ) -> dp.Forecaster:
     """Create the current forecaster if it does not exist."""
     name = model_tag.replace("-", "_")
-    # we are not using app version any more,
-    # this is stored in the forecast metadata
-    version = "2.8.0"
-
+    
     forecasters = (
         await client.list_forecasters(
             dp.ListForecastersRequest(forecaster_names_filter=[name]),
@@ -91,18 +85,18 @@ async def fetch_or_create_forecaster(
     # Forecaster does not exist, create it
     if len(forecasters) == 0:
         create_forecaster_response = await client.create_forecaster(
-            dp.CreateForecasterRequest(name=name, version=version),
+            dp.CreateForecasterRequest(name=name, version=forecast_version),
         )
         return create_forecaster_response.forecaster
 
     else:
-        filtered_forecasters = [f for f in forecasters if f.forecaster_version == version]
+        filtered_forecasters = [f for f in forecasters if f.forecaster_version == forecast_version]
 
         # Forecaster version does not exist, update it
         if len(filtered_forecasters) == 0:
 
             update_forecaster_response = await client.update_forecaster(
-                dp.UpdateForecasterRequest(name=name, new_version=version),
+                dp.UpdateForecasterRequest(name=name, new_version=forecast_version),
             )
             return update_forecaster_response.forecaster
 
@@ -159,17 +153,17 @@ async def build_multi_forecast_creation_request(
     model_tag: str,
     init_time_utc: pd.Timestamp,
     client: dp.DataPlatformDataServiceStub,
-    metadata: Struct,
+    metadata: Struct | None,
 ) -> list[dp.CreateForecastRequest]:
-    """Save forecast DataArray to data platform.
+    """Build a list of create-forecast requests for all forecasted locations.
 
     Args:
         forecast_normed_da: DataArray of normalized forecasts for all GSPs
         locations: Mapping of GSP IDs to location summaries
-        model_tag: the name of the model to saved to the database
+        model_tag: The name of the model to saved to the database
         init_time_utc: Forecast initialization time
         client: Data platform client. If None, a new client will be created.
-        metadata: Optional metadata to include with the forecast.
+        metadata: Optional metadata to assign to each location forecast.
     """
     # Fetch the forecaster and adjuster forecaster in parallel
     forecaster, adjuster_forecaster = await asyncio.gather(
@@ -201,7 +195,7 @@ async def build_multi_forecast_creation_request(
     )
 
     request = build_forecast_creation_request(
-        gsp_normed_da=da_adjusted_forecast,
+        da_forecast=da_adjusted_forecast,
         forecaster=adjuster_forecaster,
         location_uuid=locations[0].location_uuid,
         init_time_utc=init_time_utc,
@@ -214,7 +208,7 @@ async def build_multi_forecast_creation_request(
 
 
 def build_forecast_creation_request(
-    gsp_normed_da: xr.DataArray,
+    da_forecast: xr.DataArray,
     forecaster: dp.Forecaster,
     location_uuid: str,
     init_time_utc: pd.Timestamp,
@@ -223,18 +217,18 @@ def build_forecast_creation_request(
     """Build a create-forecast request from a DataArray forecast for a single location.
 
     Args:
-        gsp_normed_da: Normalized DataArray for a single GSP
+        da_forecast: Normalized DataArray for a single location
         forecaster: Forecaster object
         location_uuid: UUID of the location
         init_time_utc: Forecast initialization time
-        metadata: Optional metadata to include with the forecast.
+        metadata: Optional metadata to assign to the forecast.
     """
-    gsp_id = int(gsp_normed_da.gsp_id.values)
-    horizons_mins = gsp_normed_da.horizon_mins.values.tolist()
+    gsp_id = int(da_forecast.gsp_id.values)
+    horizons_mins = da_forecast.horizon_mins.values.tolist()
     plevels = ["p10", "p50", "p90"]
 
     forecast_array = (
-        gsp_normed_da.transpose("valid_times_utc", "output_label")
+        da_forecast.transpose("valid_times_utc", "output_label")
         .sel(output_label=plevels)
     ).values
 
@@ -268,85 +262,6 @@ def build_forecast_creation_request(
         values=forecast_value_requests,
         metadata=metadata,
     )
-
-async def fetch_adjuster_values(
-    client: dp.DataPlatformDataServiceStub,
-    location_uuid: str,
-    init_time_utc: pd.Timestamp,
-    forecaster: dp.Forecaster,
-) -> dict[int, float]:
-    """Make a forecaster adjuster based on week average deltas."""
-    deltas = (
-        await client.get_week_average_deltas(
-            dp.GetWeekAverageDeltasRequest(
-                location_uuid=location_uuid,
-                energy_source=dp.EnergySource.SOLAR,
-                pivot_timestamp_utc=convert_to_utc_datetime(init_time_utc),
-                forecaster=forecaster,
-                observer_name="pvlive_day_after",
-            ),
-        )
-    ).deltas
-
-    return {d.horizon_mins: d.delta_fraction for d in deltas}
-
-
-def apply_adjuster_values(
-    da_forecast: xr.DataArray,
-    adjuster_values: dict[int, float],
-    effective_capacity_watts: float,
-) -> xr.DataArray:
-    """Apply adjuster values to a forecast DataArray."""
-    adjuster_values_array = np.zeros(len(da_forecast.horizon_mins.values.tolist()), dtype=float)
-    for i, h in enumerate(da_forecast.horizon_mins.values.tolist()):
-        if h in adjuster_values:
-            adjuster_values_array[i] = adjuster_values[h]
-        else:
-            logger.warning(f"No adjuster value found for horizon_mins={h}; using 0.0 as default")
-
-    # Limit adjuster values to be no more than 1 GW
-    frac_1gw = 1e9 / effective_capacity_watts
-    adjuster_values_array = np.clip(adjuster_values_array, -frac_1gw, frac_1gw)
-
-    # Limit adjuster values to be no more than 10% of the forecast value
-    frac_10pc = da_forecast.sel(output_label="p50").values * 0.1
-    adjuster_values_array = np.clip(adjuster_values_array, -frac_10pc, frac_10pc)
-
-    da_adjuster_values = xr.DataArray(
-        data=adjuster_values_array,
-        dims=["valid_times_utc"],
-        coords={"valid_times_utc": da_forecast.valid_times_utc},
-    )
-
-    # Adjuster values are the average of (forecast - observed) so we need to subtract
-    # Also force the adjusted forecast to be positive by clipping at 0.0
-    ds_adjusted_forecast = (da_forecast - da_adjuster_values).clip(0, 1)
-
-    return ds_adjusted_forecast
-
-
-async def calculate_adjusted_forecast(
-    client: dp.DataPlatformDataServiceStub,
-    location: dp.ListLocationsResponseLocationSummary,
-    init_time_utc: pd.Timestamp,
-    da_forecast: xr.DataArray,
-    forecaster: dp.Forecaster,
-) -> xr.DataArray:
-    """Make an adjusted forecast based on week average deltas."""
-    adjuster_values = await fetch_adjuster_values(
-        client=client,
-        location_uuid=location.location_uuid,
-        init_time_utc=init_time_utc,
-        forecaster=forecaster,
-    )
-
-    da_adjusted_forecast = apply_adjuster_values(
-        da_forecast=da_forecast,
-        adjuster_values=adjuster_values,
-        effective_capacity_watts=location.effective_capacity_watts,
-    )
-
-    return da_adjusted_forecast
 
 
 async def write_forecasts_to_data_platform(
